backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
Pillow
pdf2image
python-dotenv
google-generativeai